Rename, reorganize
Session is no longer a namespace
API/EventLoop move under Neovim::
EventLoop is the namespace for Connection/MessageBuilder/Serializer
