+++ script/generate_docs
@@ -12,8 +12,8 @@ tabpage_docs = []
 nvim_exe = ENV.fetch("NVIM_EXECUTABLE", "nvim")
 nvim_vrs = %x(#{nvim_exe} --version).split("\n").first
 
-connection = Neovim::Session::Connection.child([nvim_exe, "-u", "NONE", "-n"])
-event_loop = Neovim::Session::EventLoop.new(connection)
+connection = Neovim::EventLoop::Connection.child([nvim_exe, "-u", "NONE", "-n"])
+event_loop = Neovim::EventLoop.new(connection)
 session = Neovim::Session.new(event_loop)
 nvim_defs = Neovim::Client.instance_methods(false)
 buffer_defs = Neovim::Buffer.instance_methods(false)
